+++ tests/unittests/image/test_tv.py
@@ -34,11 +34,11 @@ class TotalVariationTester(TotalVariation):
         super().update(img=img)
 
 
-def total_variaion_tester(preds, target, reduction="mean"):
+def _total_variaion_tester(preds, target, reduction="mean"):
     return total_variation(preds, reduction)
 
 
-def total_variation_kornia_tester(preds, target, reduction):
+def _total_variation_kornia_tester(preds, target, reduction):
     score = kornia_total_variation(preds).sum(-1)
     if reduction == "sum":
         return score.sum()
@@ -80,7 +80,7 @@ class TestTotalVariation(MetricTester):
             preds,
             target,
             TotalVariationTester,
-            partial(total_variation_kornia_tester, reduction=reduction),
+            partial(_total_variation_kornia_tester, reduction=reduction),
             metric_args={"reduction": reduction},
         )
 
@@ -89,8 +89,8 @@ class TestTotalVariation(MetricTester):
         self.run_functional_metric_test(
             preds,
             target,
-            total_variaion_tester,
-            partial(total_variation_kornia_tester, reduction=reduction),
+            _total_variaion_tester,
+            partial(_total_variation_kornia_tester, reduction=reduction),
             metric_args={"reduction": reduction},
         )
 
@@ -100,13 +100,13 @@ class TestTotalVariation(MetricTester):
             preds,
             target,
             TotalVariationTester,
-            total_variaion_tester,
+            _total_variaion_tester,
         )
 
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires cuda")
     def test_sam_half_gpu(self, preds, target, reduction):
         """Test for half precision on GPU."""
-        self.run_precision_test_gpu(preds, target, TotalVariationTester, total_variaion_tester)
+        self.run_precision_test_gpu(preds, target, TotalVariationTester, _total_variaion_tester)
 
 
 def test_correct_args():