+++ tests/unittests/regression/test_minkowski_distance.py
@@ -87,11 +87,13 @@ class TestMinkowskiDistance(MetricTester):
 
 
 def test_error_on_different_shape():
+    """Test that error is raised on different shapes of input."""
     metric = MinkowskiDistance(5.1)
     with pytest.raises(RuntimeError, match="Predictions and targets are expected to have the same shape"):
         metric(torch.randn(50), torch.randn(100))
 
 
 def test_error_on_wrong_p_arg():
+    """Test that error is raised if wrongly p argument is provided."""
     with pytest.raises(TorchMetricsUserError, match="Argument ``p`` must be a float.*"):
         MinkowskiDistance(p=-10)